feat: TeamMatchTab 팀 경기 불러오기 구현
- TeamDetailScreen에서 전환되는 TeamMatchTab 구현
- fetchTeamMatches API 연결
- 팀 경기 등록 navigation 및 API 연결 필요

S12P31A404-233

+++ frontend/mobile/src/main/java/com/ballog/mobile/data/api/MatchApi.kt
@@ -23,7 +23,7 @@ interface MatchApi {
     @GET("v1/matches/teams/{teamId}")
     suspend fun getTeamMatches(
         @Header("Authorization") token: String,
-        @Path("teamId") teamId: Long,
+        @Path("teamId") teamId: Int,
         @Query("month") month: String
     ): Response<ApiResponse<MatchListResponse>>
